4 Channels working, Filters fixed, no aux_retrigger
Channels 0 and 1 act a little off for the gamma waves. Since it might be
a dac problem, try getting rid of the if statement on the 1.8
multiplier. If gamma waves for channels 0 and 1 look off, add the if chl
== 2 || 3 back and applying the 1.8 multiplier on the front end side.
Also, the offsets will most likely need tinkering. Try to avoid messing
with the actual number for the gain multiplier.